Return array of validation message errors in UserController@store
Since the messages() method returns an array of only 1 error at a time I figured I would just return the first key right here so that I wouldn't have to deal with it on the client-side.

However, I found out that the JWT auth also returns messages in that format and since I can't change that it's better for me to keep it consistent throughout.