Add SoftDelete trait and create generateCode method on VerificationCode Lucid model
I've added the SoftDelete trait because I want to be able to call .discard() on my verifications code and easily mark them as deleted in the database so I can know when they are used and not used.

The generateCode() method is like the name suggest it, to generate a random 5 digit code to be assigned as the verification code for that entry.