Correct bug AuthController@verifyNumber not working
I spent the entire afternoon trying to figure it out but it was a silly problem.

The result from the promise is not a regular object so the rows fetched needs to be access via data.rows.

I also figured out I could just use .first() to skip all that and so I did.

I also replaced the verbose query to delete with the discard because I don't want to permantantly delete anything from my database.