Change the nature of Lucid relationship between User and WeCareContact
It can't be many to many because each user may have a different name in their phone for the same contact. If I change my sister's name in my phone , it should not change the name that my mom has set in her phone even though it's the same contact with the same number.