Add default values to timestamp tables in migrations
By default these columns are null but by setting both argument to true it will create a table where the default value is `CURRENT_TIMESTAMP` which I prefer because I don't want the fact that my timestamp be entirely dependant on the framework. The mySQL features work just fine and so they be added for an added layer of security and consistency.