Create two seperate routes for validating phone number
I need to validate the phone number for registering, and for resetting the password. Unfortunately, both of these do opposite operations, one checks that a number must be in use, the other check that a number isn't already used.

I originally to do just do simple logic to verify which case this method is running under but then I realized that during the forgot password flow, I need to to return that user's ID so that when the new password is submitted I know which user to assign it to. At that point I decided it would be messy to try to force that in there so I just separated them into 2 endpoints.