Minor changes to the validator in UserController@update
I had to add the missing .field prop on the error message because or else on the front-end I wouldn't know to which field the returned error belonged to.

I also wrote new validation messages because I don't like having the first name written as `firstname` in my error messages.