Add generateDefault method on Guardian model
I need to have default settigns for the guardians settings and I thought about doing it on the migration so that the default values are directly on the table/database but I didn't because to modify it later would be annoying because I would have to touch/mess with the table schema.

This is the cleanest option to easily tweak & modify.