Add `credit_applied_at` column to referrals table
I was in the process of figuring out how should I keep track of when a referral credit was counted as 'used' to avoid re-use of the same referrals multiple times.

I figured the most efficient and direct way would be to add a column to the `referrals` table instead of making another table like `referral_rewards` because I would then be forced of trying to keep both tables in sync which is uneeded added complexity.