Create route & method for ReadingController@dataForThisMonth
So that I can pull in all the logged in user's data for this month. I figured it would be simpler to have a single endpoint that pulled all of that months data instead of having several individuals ones for heart, and breathing, etc..

I may write those depending on my needs later but currently this and the endpoint to pull in a specifically mentioned month are all I need.