Create route & method for ReferralController@applyAsUser
To apply a user's credits because of the friends they referred to the app. Optimally, this endpoint would be called whenever the billing cycle is about to end and whenever another starts.

However as I was writing this I began to think that it might be better to take care of redeeming referral bonuses upon the billing endpoint being called from Braintree's API or something.. This was written experimentally just so that I can see how I might do but and it is very much subject to change & deletion based on what makes the most sense whenever I will implement the billing and subscription section of the server/API.